chore(helpers): update rollup to ts
