+++ helpers/types.ts
@@ -1,0 +1,6 @@
+export interface IHookDesc {
+  name: string;
+  description: string;
+  packageName?: string;
+  directoryName?: string;
+}
