+++ package.json
@@ -37,7 +37,7 @@
     "build:custom-github-actions": "cd ./.github/actions && for d in ./*/ ; do (cd \"$d\" && npm install && npm run build); done",
     "build:esm": "tsc -m esNext --outDir dist/esm",
     "build:cjs": "tsc -m commonjs --outDir dist/cjs",
-    "build:umd": "rollup -c rollup.config.js",
+    "build:umd": "rollup -c rollup.config.ts --configPlugin typescript",
     "build:types": "rollup -c dts.rollup.config.js",
     "build": "yarn build:esm && yarn build:cjs && yarn build:umd && yarn build:types",
     "generate:changelog": "gulp changelog",
